move test files to core module
