+++ addons/monitoring-addon/monitoring-core/src/test/java/org/kie/kogito/monitoring/core/integration/DaysAndTimeDurationHandlerTest.java
@@ -13,26 +13,25 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.monitoring.integration;
+package org.kie.kogito.monitoring.core.integration;
 
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
+import java.time.Duration;
 
 import io.prometheus.client.CollectorRegistry;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.kie.kogito.monitoring.system.metrics.dmnhandlers.DaysAndTimeDurationHandler;
 import org.kie.kogito.monitoring.system.metrics.dmnhandlers.DecisionConstants;
-import org.kie.kogito.monitoring.system.metrics.dmnhandlers.LocalDateTimeHandler;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class LocalDateTimeHandlerTest extends AbstractQuantilesTest<LocalDateTimeHandler> {
+public class DaysAndTimeDurationHandlerTest extends AbstractQuantilesTest<DaysAndTimeDurationHandler> {
 
     @BeforeEach
     public void setUp() {
         registry = new CollectorRegistry();
-        handler = new LocalDateTimeHandler("hello", registry);
+        handler = new DaysAndTimeDurationHandler("hello", registry);
     }
 
     @AfterEach
@@ -41,14 +40,14 @@ public class LocalDateTimeHandlerTest extends AbstractQuantilesTest<LocalDateTim
     }
 
     @Test
-    public void givenLocalDateTimeMetricsWhenMetricsAreStoredThenTheQuantilesAreCorrect() {
+    public void givenDurationMetricsWhenMetricsAreStoredThenTheQuantilesAreCorrect() {
         // Arrange
-        LocalDateTime now = LocalDateTime.now();
-        Long expectedValue = now.toInstant(ZoneOffset.UTC).toEpochMilli();
+        Integer expectedValue = 10000;
+        Duration duration = Duration.ofMillis(expectedValue);
         Double[] quantiles = new Double[]{0.1, 0.25, 0.5, 0.75, 0.9, 0.99};
 
         // Act
-        handler.record("decision", ENDPOINT_NAME, now);
+        handler.record("decision", ENDPOINT_NAME, duration);
 
         // Assert
         for (Double key : quantiles) {
