+++ addons/monitoring-addon/monitoring-core/src/test/java/org/kie/kogito/monitoring/core/integration/LocalDateTimeHandlerTest.java
@@ -13,25 +13,26 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.monitoring.integration;
+package org.kie.kogito.monitoring.core.integration;
 
-import java.time.Period;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
 
 import io.prometheus.client.CollectorRegistry;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.kie.kogito.monitoring.system.metrics.dmnhandlers.DecisionConstants;
-import org.kie.kogito.monitoring.system.metrics.dmnhandlers.YearsAndMonthsDurationHandler;
+import org.kie.kogito.monitoring.system.metrics.dmnhandlers.LocalDateTimeHandler;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class YearsAndMonthsDurationHandlerTest extends AbstractQuantilesTest<YearsAndMonthsDurationHandler> {
+public class LocalDateTimeHandlerTest extends AbstractQuantilesTest<LocalDateTimeHandler> {
 
     @BeforeEach
     public void setUp() {
         registry = new CollectorRegistry();
-        handler = new YearsAndMonthsDurationHandler("hello", registry);
+        handler = new LocalDateTimeHandler("hello", registry);
     }
 
     @AfterEach
@@ -40,14 +41,14 @@ public class YearsAndMonthsDurationHandlerTest extends AbstractQuantilesTest<Yea
     }
 
     @Test
-    public void givenYearsAndMonthsMetricsWhenMetricsAreStoredThenTheQuantilesAreCorrect() {
+    public void givenLocalDateTimeMetricsWhenMetricsAreStoredThenTheQuantilesAreCorrect() {
         // Arrange
-        Integer expectedValue = 12;
-        Period period = Period.ofMonths(expectedValue);
+        LocalDateTime now = LocalDateTime.now();
+        Long expectedValue = now.toInstant(ZoneOffset.UTC).toEpochMilli();
         Double[] quantiles = new Double[]{0.1, 0.25, 0.5, 0.75, 0.9, 0.99};
 
         // Act
-        handler.record("decision", ENDPOINT_NAME, period);
+        handler.record("decision", ENDPOINT_NAME, now);
 
         // Assert
         for (Double key : quantiles) {
@@ -55,4 +56,3 @@ public class YearsAndMonthsDurationHandlerTest extends AbstractQuantilesTest<Yea
         }
     }
 }
-
