+++ addons/monitoring-addon/monitoring-core/src/test/java/org/kie/kogito/monitoring/core/integration/LocalTimeHandlerTest.java
@@ -13,25 +13,25 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.monitoring.integration;
+package org.kie.kogito.monitoring.core.integration;
 
-import java.time.Duration;
+import java.time.LocalTime;
 
 import io.prometheus.client.CollectorRegistry;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.kie.kogito.monitoring.system.metrics.dmnhandlers.DaysAndTimeDurationHandler;
 import org.kie.kogito.monitoring.system.metrics.dmnhandlers.DecisionConstants;
+import org.kie.kogito.monitoring.system.metrics.dmnhandlers.LocalTimeHandler;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class DaysAndTimeDurationHandlerTest extends AbstractQuantilesTest<DaysAndTimeDurationHandler> {
+public class LocalTimeHandlerTest extends AbstractQuantilesTest<LocalTimeHandler> {
 
     @BeforeEach
     public void setUp() {
         registry = new CollectorRegistry();
-        handler = new DaysAndTimeDurationHandler("hello", registry);
+        handler = new LocalTimeHandler("hello", registry);
     }
 
     @AfterEach
@@ -40,14 +40,14 @@ public class DaysAndTimeDurationHandlerTest extends AbstractQuantilesTest<DaysAn
     }
 
     @Test
-    public void givenDurationMetricsWhenMetricsAreStoredThenTheQuantilesAreCorrect() {
+    public void givenLocalTimeMetricsWhenMetricsAreStoredThenTheQuantilesAreCorrect() {
         // Arrange
-        Integer expectedValue = 10000;
-        Duration duration = Duration.ofMillis(expectedValue);
+        LocalTime now = LocalTime.now();
+        int expectedValue = now.toSecondOfDay();
         Double[] quantiles = new Double[]{0.1, 0.25, 0.5, 0.75, 0.9, 0.99};
 
         // Act
-        handler.record("decision", ENDPOINT_NAME, duration);
+        handler.record("decision", ENDPOINT_NAME, now);
 
         // Assert
         for (Double key : quantiles) {
