+++ addons/monitoring-addon/monitoring-core/src/test/java/org/kie/kogito/monitoring/core/mock/DMNDecisionResultMock.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.monitoring.mocks;
+package org.kie.kogito.monitoring.core.mock;
 
 import java.util.ArrayList;
 import java.util.List;
